Add EObject helper methods to SearchUtils
findChildWithName
Returns an Optional containing a Term that is a child of the given taxonomy EObject with the given name if it exists

getChildren
Returns children of a taxonomy EObject. Dimension if model, subclasses if Term. Raises an IllegalArgumentException if an unrecognised taxonomy element is given

childWithNameExists
checks wether the given taxonomy element has a child with the given name